+++ tests/test_updater.py
@@ -246,8 +246,9 @@ class LocalUpdaterTest(unittest.TestCase):
             assert params['DistributionId'] == "DWIGHTENIS"
             assert params['InvalidationBatch']['CallerReference'].startswith('tz_1234-')
             assert params['InvalidationBatch']['Paths'] == {
-                'Quantity': 1,
-                'Items': ['/v1/buckets/destbucket/collections/destcollection*']
+                'Quantity': 2,
+                'Items': ['/v1/buckets/destbucket/collections/destcollection*',
+                          '/v1/buckets/monitor/collections/changes*']
             }
 
     def test_does_not_fail_when_cache_invalidation_does(self):